Move create_from_attributes from Organisation model to another Importers class
